+++ app/models/membership.rb
@@ -1,0 +1,9 @@
+class Membership < ApplicationRecord
+  include Discardable
+
+  belongs_to :user
+  belongs_to :workspace
+  belongs_to :role
+
+  validates :user_id, uniqueness: { scope: :workspace_id }
+end