+++ spec/factories/workspaces.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :workspace do
+    name { Faker::Company.name }
+    plan { "free" }
+  end
+end